fix: 处理 PR #7 Copilot 审核意见
- upolynomial.hh: poly_convert(Zp→ZZ) 末尾加 normalization()
- number.hh: __nmod_mul 加 assert(_p != 0) 防御
- test_mtshl_m1.cc: get_coeff 返回类型 uint32_t → uint64_t
- test_number.cc: 测试名改为下划线风格

+++ test/test_mtshl_m1.cc
@@ -6,7 +6,7 @@ using namespace clpoly;
 // ================================================================
 // 辅助：从 UPZp 中按次数提取系数
 // ================================================================
-static uint32_t get_coeff(const upolynomial_<Zp>& p, int64_t deg)
+static uint64_t get_coeff(const upolynomial_<Zp>& p, int64_t deg)
 {
     for (const auto& term : p)
         if (term.first.deg() == deg)
